+++ spec/support/commands/validation_token_command.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class ValidationTokenCommand < ApplicationCommand
+
+  def call
+    if ctx.simulate_token_collision
+      fail!("Validation token already exists")
+    else
+      context.validation_token = "123abc-456def"
+    end
+  end
+
+end